Consolidate Document tests into a single file
